Gracefully handle HTTP errors in SimSimi.com
Any HTTP error thrown in SimSimi.com will cause the application to crash. This will result in Telegram hitting our Google App Engine application continuously until it gets a response (causing the application to start up and crash over and over again, as long as there's a problem with the SimSimi site). To mitigate this "boot loop" and satiate the hungry Telegram, we will simply return the stringified message of the HTTP error as a bot response.